reduce float size

+++ compute/cupy/grayscott_cupy.py
@@ -31,7 +31,7 @@ def grayscott_cupy(U, V, Du, Dv, F, k, delta_t, nb_frame, step_frame):
     :param step_frame: copy frame each 'step_frame' finit difference iteration
     '''
     n_x, n_y = U.shape[0], U.shape[1]
-    frames_V = np.empty((nb_frame, n_x, n_y), dtype=V.dtype)
+    frames_V = np.empty((nb_frame, n_x, n_y), dtype=np.float16)
     stencil = np.array([[0, 1, 0],
                         [1, -4, 1],
                         [0, 1, 0]],
@@ -46,7 +46,7 @@ def grayscott_cupy(U, V, Du, Dv, F, k, delta_t, nb_frame, step_frame):
             Lu = conv2d_gpu(u_gpu, stl_gpu, 'same', 'fill', 0)
             Lv = conv2d_gpu(v_gpu, stl_gpu, 'same', 'fill', 0)
             grayscott_kernel(Lu, Lv, u_gpu, v_gpu, Du, Dv, F, k, delta_t)
-        frames_V[idx_fr ,:,:] = v_gpu.get()
+        frames_V[idx_fr ,:,:] = v_gpu.get().astype(np.float16)
     return frames_V
